// src/payments/enums/payment-event-type.enum.ts
export enum PaymentEventType {
  PAYMENT_INITIATED = 'payment.initiated',
  PAYMENT_SUCCESSFUL = 'payment.successful',
  PAYMENT_FAILED = 'payment.failed',
  REFUND_INITIATED = 'refund.initiated',
  REFUND_SUCCESSFUL = 'refund.successful',
  REFUND_FAILED = 'refund.failed',
  TRANSFER_SUCCESSFUL = 'transfer.successful',
  TRANSFER_FAILED = 'transfer.failed',
}